// src/utils/enums/user.enum.ts
export enum UserRole {
  CUSTOMER = 'Customer',
  REGULAR = 'Regular',
  ADMIN = 'Admin',
  SUPER_ADMIN = 'Super Admin',
  PROJECT_MANAGER = 'Project Manager',
  CUSTOMER_SUPPORT = 'Customer Support',
  SETTLEMENT = 'Settlement',
  GAME_MANAGER = 'Game Manager',
  MARKETING_ADMIN = 'Marketing Admin',
  ADVISOR = 'Advisor',
}